pyproject.toml
[tool.poetry]
name = "shopapi"
version = "0.1.0"
description = ""
authors = ["Tomas Votava <info@tomasvotava.eu>"]

[mypy]
ignore_missing_imports = true
show_error_codes = true
show_column_numbers = true

[tool.black]
line-length = 119

[tool.poe.tasks]
dev = "uvicorn --host localhost --port 8000 --reload --log-level=debug main:app"
start = "uvicorn --host localhost --port 8000 main:app"
db-init-migrations = "aerich init -t shopapi.schemas.models.TORTOISE_ORM"
db-init = "aerich init-db"
docs = "bash build-docs.sh"
generate-secure-key = "openssl rand -hex 32"
pylint = "pylint main.py shopapi"
mypy = "mypy --config-file pyproject.toml main.py shopapi/"
lint = "sh -c 'poe pylint; poe mypy'"

[tool.poetry.dependencies]
python = "^3.9"
tortoise-orm = "^0.16.21"
pydantic = {extras = ["email"], version = "^1.8.1"}
fastapi = "^0.63.0"
aerich = "^0.5.0"
python-dotenv = "^0.15.0"
config-proxy = "^0.1.5"
fastapi-sso = "^0.2.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
asyncpg = "^0.22.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.7.2"
poethepoet = "^0.10.0"
uvicorn = "^0.13.4"
pdoc3 = "^0.9.2"
mypy = "^0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
